Reduce code duplication
