+++ plutus-core/plutus-core/src/PlutusCore/Default/Universe.hs
@@ -40,6 +40,8 @@ module PlutusCore.Default.Universe
   , pattern DefaultUniPair
   , defaultUniSize
   , noMoreTypeFunctions
+  , caseBuiltinDefault
+  , caseBuiltinNoData
   , module Export -- Re-exporting universes infrastructure for convenience.
   ) where
 
@@ -916,47 +918,79 @@ instance AnnotateCaseBuiltin DefaultUni where
         _ -> Left "Casing on pair requires exactly one branch"
     _ -> Left $ display (void ty) <> " isn't supported in 'case'"
 
+-- | Case on a 'DefaultUni' value, delegating casing on 'Data' to the supplied handler.
+caseBuiltinDefault
+  :: ( Data
+       -> Some (ValueOf DefaultUni)
+       -> Vector.Vector term
+       -> HeadSpine Text term (Some (ValueOf DefaultUni))
+     )
+  -> Some (ValueOf DefaultUni)
+  -> Vector.Vector term
+  -> HeadSpine Text term (Some (ValueOf DefaultUni))
+caseBuiltinDefault caseData someVal@(Some (ValueOf uni x)) branches = case uni of
+  DefaultUniUnit
+    | 1 == len -> HeadOnly $ branches Vector.! 0
+    | otherwise -> HeadError $ outOfBoundsErr someVal branches
+  DefaultUniBool -> case x of
+    -- We allow there to be only one branch as long as the scrutinee is 'False'.
+    -- This is strictly to save size by not having the 'True' branch if it was gonna be
+    -- 'Error' anyway.
+    False | len == 1 || len == 2 -> HeadOnly $ branches Vector.! 0
+    True | len == 2 -> HeadOnly $ branches Vector.! 1
+    _ -> HeadError $ outOfBoundsErr someVal branches
+  DefaultUniInteger
+    | 0 <= x && x < toInteger len -> HeadOnly $ branches Vector.! fromInteger x
+    | otherwise -> HeadError $ outOfBoundsErr someVal branches
+  DefaultUniData -> caseData x someVal branches
+  DefaultUniList ty
+    | len == 1 ->
+        case x of
+          [] -> HeadError "Expected non-empty list, got empty list for casing list"
+          (y : ys) -> headSpine (branches Vector.! 0) [someValueOf ty y, someValueOf uni ys]
+    | len == 2 ->
+        case x of
+          [] -> HeadOnly $ branches Vector.! 1
+          (y : ys) -> headSpine (branches Vector.! 0) [someValueOf ty y, someValueOf uni ys]
+    | otherwise -> HeadError $ outOfBoundsErr someVal branches
+  DefaultUniPair tyL tyR
+    | len == 1 ->
+        case x of
+          (l, r) -> headSpine (branches Vector.! 0) [someValueOf tyL l, someValueOf tyR r]
+    | otherwise -> HeadError $ outOfBoundsErr someVal branches
+  _ -> HeadError $ display uni <> " isn't supported in 'case'"
+  where
+    !len = Vector.length branches
+{-# INLINE caseBuiltinDefault #-}
+
+-- | The built-in caser used when casing is available except on 'Data'.
+caseBuiltinNoData
+  :: Some (ValueOf DefaultUni)
+  -> Vector.Vector term
+  -> HeadSpine Text term (Some (ValueOf DefaultUni))
+caseBuiltinNoData =
+  caseBuiltinDefault $ \_ _ _ -> HeadError "Casing on data is not supported"
+{-# INLINE caseBuiltinNoData #-}
+
+caseBuiltinData
+  :: Data
+  -> Some (ValueOf DefaultUni)
+  -> Vector.Vector term
+  -> HeadSpine Text term (Some (ValueOf DefaultUni))
+caseBuiltinData x someVal branches = case x of
+  Data.Constr tag fields
+    | 0 <= tag && tag < toInteger len ->
+        headSpine
+          (branches Vector.! fromInteger tag)
+          [someValueOf (DefaultUniList DefaultUniData) fields]
+    | otherwise -> HeadError $ outOfBoundsErr someVal branches
+  _ -> HeadError "Casing on data only supports Data.Constr values"
+  where
+    !len = Vector.length branches
+{-# INLINE caseBuiltinData #-}
+
 instance CaseBuiltin DefaultUni where
-  caseBuiltin someVal@(Some (ValueOf uni x)) branches = case uni of
-    DefaultUniUnit
-      | 1 == len -> HeadOnly $ branches Vector.! 0
-      | otherwise -> HeadError $ outOfBoundsErr someVal branches
-    DefaultUniBool -> case x of
-      -- We allow there to be only one branch as long as the scrutinee is 'False'.
-      -- This is strictly to save size by not having the 'True' branch if it was gonna be
-      -- 'Error' anyway.
-      False | len == 1 || len == 2 -> HeadOnly $ branches Vector.! 0
-      True | len == 2 -> HeadOnly $ branches Vector.! 1
-      _ -> HeadError $ outOfBoundsErr someVal branches
-    DefaultUniInteger
-      | 0 <= x && x < toInteger len -> HeadOnly $ branches Vector.! fromInteger x
-      | otherwise -> HeadError $ outOfBoundsErr someVal branches
-    DefaultUniData -> case x of
-      Data.Constr tag fields
-        | 0 <= tag && tag < toInteger len ->
-            headSpine
-              (branches Vector.! fromInteger tag)
-              [someValueOf (DefaultUniList DefaultUniData) fields]
-        | otherwise -> HeadError $ outOfBoundsErr someVal branches
-      _ -> HeadError "Casing on data only supports Data.Constr values"
-    DefaultUniList ty
-      | len == 1 ->
-          case x of
-            [] -> HeadError "Expected non-empty list, got empty list for casing list"
-            (y : ys) -> headSpine (branches Vector.! 0) [someValueOf ty y, someValueOf uni ys]
-      | len == 2 ->
-          case x of
-            [] -> HeadOnly $ branches Vector.! 1
-            (y : ys) -> headSpine (branches Vector.! 0) [someValueOf ty y, someValueOf uni ys]
-      | otherwise -> HeadError $ outOfBoundsErr someVal branches
-    DefaultUniPair tyL tyR
-      | len == 1 ->
-          case x of
-            (l, r) -> headSpine (branches Vector.! 0) [someValueOf tyL l, someValueOf tyR r]
-      | otherwise -> HeadError $ outOfBoundsErr someVal branches
-    _ -> HeadError $ display uni <> " isn't supported in 'case'"
-    where
-      !len = Vector.length branches
+  caseBuiltin = caseBuiltinDefault caseBuiltinData
   {-# INLINE caseBuiltin #-}
 
 {- Note [Stable encoding of tags]
